% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channel_width.R
\name{channel_width}
\alias{channel_width}
\title{Estimates the width of the active channel}
\usage{
channel_width(c_line, dgo, water_mask, max_width = 15000)
}
\arguments{
\item{c_line}{centerline object derived with the centerline() function}

\item{dgo}{an sf object with the disaggregated river corridor derived with the disaggregation() function}

\item{water_mask}{SpatRaster or file path to a binary raster representing an approximation of bankfull discharge;
in a simple case, this raster can be obtained from thresholding over the JRC global surface water dataset (https://global-surface-water.appspot.com)}

\item{max_width}{estimated maximum width of the river corridor; in doubt increase, it is clipped to the dgo outline}
}
\value{
a data.frame with two columns, the segment id of the dgo object and the active channel width (acw)
}
\description{
The centerline function derives the centerline of the river corridor mask
}
\author{
Florian Betz
}
