% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densify_sf.R
\name{densify_sf}
\alias{densify_sf}
\title{Helper function to densify points on the boundary of a polygon}
\usage{
densify_sf(geometry, distance)
}
\arguments{
\item{geometry}{Input sf or sfc object whose outline vertices are densified}

\item{distance}{Interval at which the new vertices are set}
}
\value{
a sfc object with the densified vertices
}
\description{
This function is used to densify the vertices on a linestring. This is used for centerline creation
}
\author{
Florian Betz
}
