% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disaggretation.R
\name{disaggregate}
\alias{disaggregate}
\title{Disaggregate the river corridor}
\usage{
disaggregate(centerline, segment_length = 100, max_width)
}
\arguments{
\item{centerline}{sf line object representing the centerline of the river corridor; it is assumed to be in projected coordinates with metric units}

\item{segment_length}{spacing of the segments for the disaggregated river corridor, defaults to 100 m}

\item{max_width}{estimation of the maximum width of the river corridor to ensure that the polygons cover the entire corridor}
}
\value{
returns a sf polygon object
}
\description{
uses centerline of the river corridor and Thiessen polygons to create disaggregated river segments
}
\references{
Alber, A., Piégay, H. (2011): Spatial disaggregation and aggregation procedures for characterizing
fluvial features at the network scale: applications to the Rhône basin (France). Geomorphology 125, 343-360.
Betz, F., Lauermann, M., Cyffka, B. (2020): Open source riverscapes:...
}
\author{
Florian Betz
}
