% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend_line.R
\name{extend_line}
\alias{extend_line}
\title{This is a helper function for constructing the centerline. Its intended for internal use only}
\usage{
extend_line(line, dist)
}
\arguments{
\item{line}{sf object containing the line to be extended}

\item{dist}{the distance in map units to extend the line}
}
\value{
an sf object containing the extended line
}
\description{
This function extends a line to both ends in the direction of the last line segment by a specified distance
The function assumes the line to be in projected coordinates
}
\author{
Florian Betz
}
