% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_channels_saga.R
\name{extract_channels_saga}
\alias{extract_channels_saga}
\title{Extracting a channel network from a digital elevation network using SAGA GIS}
\usage{
extract_channels_saga(
  dem,
  preprocess = "breach",
  min_slope = 1e-04,
  initiation = "cit",
  threshold,
  min_seglength = 10,
  saga_path,
  out_dir = tempdir()
)
}
\arguments{
\item{dem}{Path to a digital elevation model stored in a common raster format}

\item{preprocess}{option to use for preprocessing the DEM in order to create a depressionless elevation model
can be "fill_sink" which is using the Wang and Liu sink filling algorithm or "breach"
which is the SAGA implementation of breaching as in the Whitebox tools}

\item{min_slope}{minimum slope to maintain when carrying out sink filling; default is 0.1}

\item{initiation}{method for channel initiation, either "flow_acc" using flow accumulation or
"cit" using the channel initiation index}

\item{threshold}{threshold used for channel initiation. Commonly, flow accumulation is used for this.}

\item{min_seglength}{length (in cells) which a channel segment needs to have at least to be returned}

\item{saga_path}{path where the SAGA installation is found (see RSAGA::rsaga.env() for details)}

\item{out_dir}{directory where to store output files}
}
\value{
a list with a raster (terra rast() format) and a vector (terra vect() format) dataset of the channel network
}
\description{
This function provides a wrapper for two different approaches for channel network delineation from a digital elevation model
}
\author{
Florian Betz
}
