% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_mode.R
\name{find_mode}
\alias{find_mode}
\title{Little helper function to find the mode in a character vector}
\usage{
find_mode(x)
}
\arguments{
\item{x}{character vector}
}
\value{
single character with the most frequent entrance in x
}
\description{
This function finds the entrance with the highest frequency in a character vector, i.e. the mode
}
\author{
Florian Betz
}
