% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzzy_overlay.R
\name{fuzzy_overlay}
\alias{fuzzy_overlay}
\title{Overlay several fuzzified indices}
\usage{
fuzzy_overlay(x, mode = "AND")
}
\arguments{
\item{x}{list of fuzzified indices}

\item{mode}{mode of overlay, either "AND" or "OR"}
}
\value{
spatRaster of a value range from 0-1 where 1 is fully connected to the channel and 0 is not connected at all.
}
\description{
Combining several fuzzified indices using defined Zadeh's classic AND/OR rules
}
\references{
Betz, F., Lauermann, M., Cyffka, B. (2018):
}
\author{
Florian Betz
}
