% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_grass.R
\name{install_grass}
\alias{install_grass}
\title{Helper function to install GRASS GIS on the machine}
\usage{
install_grass(
  url =
    "https://grass.osgeo.org/grass84/binary/mswindows/native/WinGRASS-8.4.0-1-Setup.exe",
  download_folder = "C:/Downloads"
)
}
\arguments{
\item{url}{download url of GRASS GIS; it defaults to GRASS 8.4.0, the latest version at the time of writing this function}

\item{download_folder}{folder where to download the GRASS installer file, typically your download folder}
}
\value{
This function doesn't return any object
}
\description{
This function downloads the GRASS GIS installer from the web and initiates the installation
}
\author{
Florian Betz
}
