% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_saga.R
\name{install_saga}
\alias{install_saga}
\title{Helper function to install SAGA GIS on the machine}
\usage{
install_saga(
  url =
    "https://sourceforge.net/projects/saga-gis/files/SAGA\%20-\%209/SAGA\%20-\%209.0.1/saga-9.0.1_x64.zip/download",
  install_folder = "C:/SAGA"
)
}
\arguments{
\item{url}{download url of SAGA GIS; it defaults to SAGA 9.0.1 as this is the version riverMapper has been tested with}

\item{install_folder}{directory, where SAGA will be living. This is not necessarily C:/Program Files}
}
\value{
This function doesn't return any object
}
\description{
This function downloads the SAGA GIS installer from the web and initiates the installation
}
\author{
Florian Betz
}
