% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_segments.R
\name{link_segments}
\alias{link_segments}
\title{Linking channel attributes and floodplain segments derived from the disaggregation function.}
\usage{
link_segments(riverline, segs)
}
\arguments{
\item{riverline}{sf object containing the riverline}

\item{segs}{sf object containing the floodplain segments as derived with the disaggregation function}
}
\value{
an sf object containing original floodplain segments with the distances along the channel
}
\description{
The link_segments() function takes the distances along the channel and assigns them to the disaggregated
floodplain segments
}
\author{
Florian Betz
}
