% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_profile.R
\name{make_profile}
\alias{make_profile}
\title{Creates a longitudinal profile using a DEM (or other raster) and a channel line as input}
\usage{
make_profile(
  channel_line,
  rast_file,
  dem,
  start = "highest",
  maxDist = 10,
  plot = TRUE
)
}
\arguments{
\item{channel_line}{sf object containing the channel line}

\item{rast_file}{terra object with the raster values to use for the profile}

\item{dem}{digital elevation model used to specify the direction of the profile}

\item{start}{where should the profile start? "Highest" will start compute the distances from source to mouth,
"lowest" from mouth to source
#'@param maxDist maximum distance (in meter) between two consecutive vertices when densifying the vertices of the line.
Default is 10 m.In general, maxDist should be in a similar range as the resolution of the underlying elevation data.}

\item{plot}{If TRUE, a line plot of the profile will be returned}
}
\value{
a data.frame with the coordinates, the profile distance and the raster value
}
\description{
This function creates longitudinal profiles from a raster (most likely a DEM) and a channel line.
It uses the profile_dist() function to compute distances along the line, extracts the raster values
and adds them to a data.frame with the coordinates and cumulative distances along the line as well as the raster values
}
\author{
Florian Betz
}
