% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ortho_line.R
\name{ortho_line}
\alias{ortho_line}
\title{Creates an orthogonal line on another line}
\usage{
ortho_line(line, line_length)
}
\arguments{
\item{line}{sf object of the line used to construct the orthogonal line on}

\item{line_length}{the length of the orthogonal line}
}
\value{
an sf object of the orthogonal line
}
\description{
This function creates an orthogonal line on another line and is designed to work as a
helper function in the channel and corridor width functions
}
\author{
Florian Betz
}
