% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pettitt_test.R
\name{pettitt_test}
\alias{pettitt_test}
\title{Test for changepoints in an univariate series using an iterative pettitt test}
\usage{
pettitt_test(data, alpha = 0.05)
}
\arguments{
\item{data}{vector of continuous data to be segmented}

\item{alpha}{significance level to be used in the pettitt test}
}
\value{
vector of changepoint locations in the input data
}
\description{
This function applies the commonly used pettitt test in an iterative way, where new segments
are created by a changepoint and searched for new changepoints until no significant result of the pettitt
test is returned anymore.
}
\author{
Florian Betz
}
