% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riverMapper.R
\docType{package}
\name{riverMapper}
\alias{riverMapper-package}
\alias{riverMapper}
\title{riverMapper}
\description{
R Package designed for mapping various geomorphological and (simple)
vegetation attributes of river corridor. The package is specifically tailored to the
needs of data-scarce environments and the use of globally available (satellite) data.
}
\details{
R-Package for River Corridor Mapping
}
\references{
Betz, F., Lauermann, M., Cyffka, B. (2020): Open source riverscapes:
}
\seealso{
Useful links:
\itemize{
  \item \url{https://fbetz-geo.github.io/riverMapper/}
}

}
\author{
Florian Betz
}
