% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation.R
\name{aggregation}
\alias{aggregation}
\title{Aggregation of the disaggregated units to uniform reaches using statistical methods}
\usage{
aggregation(river_df, method = "bcp", thresh)
}
\arguments{
\item{river_df}{data.frame with the river corridor attributes on the disaggregated level.}

\item{method}{can be one of the following: pettitt for the univariate pettitt test, cluster for clustering or
bcp for Bayesian changepoint estimation; default is "bcp"}

\item{thresh}{threshold for the changepoint probability to be used by the bcp method}
}
\description{
This function provides various methods to aggregate the disaggregated river corridor into homogeneous reaches.
It provides the univariate Pettitt test, a multivariate Bayesian method as well as a multivariate clustering method
}
