% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centerline.R
\name{centerline}
\alias{centerline}
\title{creates a centerline from the raster based corridor mask}
\usage{
centerline(mask, smooth)
}
\arguments{
\item{mask}{spatRaster of the binary river corridor mask where the river corridor is encoded with 1 and non-corridor is NA}

\item{smooth}{smoothing factor for converting the raster mask to polygon}
}
\value{
a list with 1) the centerline as sf line object and 2) the smoothed polygon as sf object
}
\description{
The centerline function derives the centerline of the river corridor mask by
}
\author{
Florian Betz
}
