% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeIndices.R
\name{computeIndices}
\alias{computeIndices}
\title{Compute Morphometric Indices}
\usage{
computeIndices(
  dem,
  channels,
  gis_backend = "saga",
  gis_path = c("vdist", "PathDist", "mrvbf"),
  index,
  out_dir = tempdir(),
  overwrite = FALSE
)
}
\arguments{
\item{dem}{path to the digital elevation model}

\item{channels}{path to raster of channel network delineated from digital elevation models}

\item{gis_backend}{which GIS software to use for computing the more complex terrain parameters;
currently only SAGA GIS is supported with GRASS GIS being available soon}

\item{gis_path}{path where the SAGA installation is found (see RSAGA::rsaga.env() for details)}

\item{index}{character; indices to be computed, can be "hdist" for horizontal distance, "vdist"
for vertical distance, "mrvbf" for multiresolution valleybottom flatness index, "TWI" for Wetness Index
or "PathDist" for Path Distance}

\item{out_dir}{directory where to store output files}

\item{overwrite}{should already existing files be overwritten? Defaults to true and only applies to the pathDistance!}
}
\value{
spatRaster with the morphometric indices as layers
}
\description{
Computes different morphometric indices useful for the fuzzy delineation of river corridors
}
\details{
Vertical Distance, and horizontal distance are currently computed using SAGA GIS. The computation is efficient, but not memory safe,
#' i.e. the entire DEM has to fit in RAM. Thus, it is recommended to watch memory use while computing these indices.
}
\references{
Betz, F., Lauermann, M., Cyffka, B. (2018): Delineation of riparian zone...
}
\author{
Florian Betz
}
