% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile_distance.R
\name{profile_distance}
\alias{profile_distance}
\title{Computing distance along a channel}
\usage{
profile_distance(riverline, maxDist = 10)
}
\arguments{
\item{riverline}{sf object containing the channel line}

\item{maxDist}{maximum distance (in meter) between two consecutive vertices when densifying the vertices of the line.
Default is 10 m.In general, maxDist should be in a similar range as the resolution of the underlying elevation data.}
}
\value{
an sf points object containing the downstream distances along the channel line along with their downstream distance
}
\description{
This function computes the distance along a channel line, where the line is "exploded" before computing distances. The function
assumes the channel line being in projected coordinates and will give an error if this is not the case.
}
\author{
Florian Betz
}
